wiregrid_kikusui: Move wg kikusui agent into package
