+++ docker-compose.yml
@@ -104,15 +104,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # Wiregrid KIKUSUI power supply agent
-  # --------------------------------------------------------------------------
-  ocs-wgkikusui-agent:
-    image: "ocs-wgkikusui-agent"
-    build: ./agents/wiregrid_kikusui/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # Wiregrid actuator agent
   # --------------------------------------------------------------------------
   ocs-wgactuator-agent:
