+++ docs/agents/wiregrid_kikusui.rst
@@ -14,9 +14,9 @@ via RS-232 (D-sub 9pin cable).
 The agent communicates with the converter via Ethernet.
 
 .. argparse::
-   :filename: ../agents/wiregrid_kikusui/kikusui_agent.py
+   :filename: ../socs/agents/wiregrid_kikusui/agent.py
    :func: make_parser
-   :prog: python3 kikusui_agent.py
+   :prog: python3 agent.py
 
 Configuration File Examples
 ---------------------------
@@ -49,8 +49,7 @@ Docker Compose
 An example docker-compose configuration::
 
     ocs-wgkikusui-agent:
-      image: simonsobs/ocs-wgkikusui-agent:latest
-      restart: always
+      image: simonsobs/socs:latest
       hostname: ocs-docker
       network_mode: "host"
       depends_on:
@@ -107,7 +106,7 @@ However, if you want to rotate the wire-grid continuousely, you can use the foll
 Agent API
 ---------
 
-.. autoclass:: agents.wiregrid_kikusui.kikusui_agent.WiregridKikusuiAgent
+.. autoclass:: socs.agents.wiregrid_kikusui.agent.WiregridKikusuiAgent
     :members:
