+++ socs/agents/ocs_plugin_so.py
@@ -24,5 +24,6 @@ for n, f in [
         ('ScpiPsuAgent', 'scpi_psu/agent.py'),
         ('TektronixAWGAgent', 'tektronix3021c/agent.py'),
         ('VantagePro2Agent', 'vantagepro2/agent.py'),
+        ('WiregridKikusuiAgent', 'wiregrid_kikusui/agent.py'),
 ]:
     ocs.site_config.register_agent_class(n, os.path.join(root, f))
